revise the interface with tough-rate
There is a race condition, where tough-rate will set session.winner after the bridge command returns (so probbaly after some kind of command-execute-complete notification from FreeSwitch).
However this will probably be much later than when `cdr-report` is sent by huge-play/middleware/cdr, since it only depends on channel-hangup-complete.
So we use the last gateway tried by tough-rate and assume it is the one we need to use, if no `session.winner` is present.