prevent replication of unused tables
The plans listed in `rating[]` are the ones the client agreed to on the date indicated (or the application date for a given plan).

There's still the issue of dynamic changes in these documents (what if nobody has signed up for the new tables at startup time?) and additions of new tables.