provide `emergency` and `onnet` dynamically
The fields are populated by `huge-play/middleware/cdr` at the time the call
is disconnected in `cdr_report`, but we never use it.
In any case, for intermediate CDRs, we must provide them.

The issue is that the fields will not be correct the very first time the
script is called.
In-call scripts should say

    if called_emergency() is true

or

    unless called_emergency() isnt false